requirements.txt
fastapi==0.115.4
graphviz==0.20.3
langchain==0.3.4
langchain_community==0.3.3
langchain_core==0.3.12
langchain_openai==0.2.3
langgraph==0.2.22
openai==1.53.0
passlib==1.7.4
pyautogen==0.3.1
pydantic==2.9.2
pydantic_settings==2.6.0
PyJWT==2.8.0
SQLAlchemy==2.0.36
starlette==0.41.2
uvicorn==0.32.0
zhipuai==2.0.1
urllib3==1.26.6
python-multipart==0.0.9
aiomysql==0.2.0
mysql-connector-python==8.3.0
email-validator==2.1.1
bcrypt==4.1.2
websockets==12.0
tencentcloud-sdk-python
python-dotenv==1.0.1
pytz==2024.2
Requests==2.32.3
greenlet==3.0.3
chromadb==0.5.18
docx2txt==0.8
pytest==8.3.3
pytest_asyncio==0.24.0
httpx==0.27.0
cos-python-sdk-v5==1.9.33
alembic==1.14.0